requirements.txt
asgiref==3.7.2
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-auditlog==2.3.0
django-auth-adfs==1.11.6
django-cleanup==8.0.0
django-debug-toolbar==4.2.0
django-jsoneditor-widget==1.3.2
django-model-utils==4.3.1
django-modeladmin-reorder==0.3.1
-e git+https://github.com/Corvia/django-tenant-users.git@6b4cb27b12b883f99fe4b0ffb6d83f3914b70a0a#egg=django_tenant_users
-e git+https://github.com/django-tenants/django-tenants.git@04164c2d38b03d3508ec109a39989d42b9262f96#egg=django_tenants
django-tinymce==3.6.1
idna==3.4
isodate==0.6.1
msal==1.25.0
msal-extensions==1.0.0
Pillow==10.1.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
